# sql: avoid ResolveBlankPaddedChar in writeTextString