writeTextString no longer uses ResolveBlankPaddedChar which was causing an increase in heap allocations mainly due to it's use of `make([]byte, t.Width())`.  Additionally in this PR I removed the call to ResolveBlankPaddedChar from writeBinaryString and other occurences in types.go.

These are the results that we now see from benchmark tests:

Micro Benchmark:
```
                      │   sec/op    │   sec/op     vs base               │
WriteTextColumnarChar   41.49µ ± 2%   24.05µ ± 3%  -42.03% (p=0.001 n=7)

                      │  before.txt  │              after.txt              │
                      │     B/op     │    B/op      vs base                │
WriteTextColumnarChar   16.00Ki ± 0%   0.00Ki ± 0%  -100.00% (p=0.001 n=7)

                      │ before.txt  │              after.txt              │
                      │  allocs/op  │  allocs/op   vs base                │
WriteTextColumnarChar   1.024k ± 0%   0.000k ± 0%  -100.00% (p=0.001 n=7)
```

Sysbench:
```
                                        │ sys_gce_before.txt │         sys_gce_after.txt         │
                                        │       sec/op       │   sec/op     vs base              │
Sysbench/SQL/1node_local/oltp_read_only          7.313m ± 4%   7.071m ± 4%  -3.32% (p=0.009 n=6)

                                        │ sys_gce_before.txt │         sys_gce_after.txt          │
                                        │        B/op        │     B/op      vs base              │
Sysbench/SQL/1node_local/oltp_read_only         972.3Ki ± 2%   895.3Ki ± 1%  -7.92% (p=0.002 n=6)

                                        │ sys_gce_before.txt │         sys_gce_after.txt          │
                                        │     allocs/op      │  allocs/op   vs base               │
Sysbench/SQL/1node_local/oltp_read_only          6.179k ± 4%   5.250k ± 2%  -15.03% (p=0.002 n=6)
```

Epic: CRDB-42584
Fixes: #134336
Release note: None